# Significantly reduce memory use of maps.Flatten