This patch addresses an issue where maps.Flatten would use an unreasonable amount of memory when being applied to large nested structures.

See benchmark comparison:

```
benchmark               old ns/op     new ns/op     delta
BenchmarkFlatten-16     9225          3773          -59.10%

benchmark               old allocs     new allocs     delta
BenchmarkFlatten-16     63             27             -57.14%

benchmark               old bytes     new bytes     delta
BenchmarkFlatten-16     9029          2405          -73.36%
```